refactor(settings): split the appearance tab into its own file
views_settings.rs sat at exactly the repository's 1,000-line ceiling,
so any further settings work had to start by moving something out.
The Appearance tab is self-contained: it reads only theme state and
emits only theme messages, so it moves whole with no new plumbing.

The one non-mechanical edit replaces an em dash in the "no themes
found" hint with a colon, since em dashes are banned in this codebase.
